Add logger interface implementation test
This test checks if the scoped logger interface is implemented
by a given module.

So far, the argument parser, and the render engine modules are
hooked up to the test.

The test itself, is simply the function name resolution during linking.
No binary execution takes place.
If the linking succeeds, the test succeeds.